+++ .eslintignore
@@ -4,7 +4,7 @@
 # The below mentioned json files and directories are purposefully
 # made invalid in order to execute the unit tests, thus thse are
 # required to be ignored by the json validator.
-/host-bmc/test/host_effecter_jsons/malformed/*
-/host-bmc/test/host_effecter_jsons/no_json/*
+/platform-mc/test/host_effecter_jsons/malformed/*
+/platform-mc/test/host_effecter_jsons/no_json/*
 /libpldmresponder/test/pdr_jsons/state_effecter/malformed/*
 /libpldmresponder/test/pdr_jsons/state_sensor/malformed/*
