+++ .linter-ignore
@@ -1,4 +1,4 @@
-host-bmc/test/host_effecter_jsons/malformed/dbus_to_terminus_effecter.json
-host-bmc/test/host_effecter_jsons/no_json/dummy.json
+platform-mc/test/host_effecter_jsons/malformed/dbus_to_terminus_effecter.json
+platform-mc/test/host_effecter_jsons/no_json/dummy.json
 libpldmresponder/test/pdr_jsons/state_effecter/malformed/effecter_pdr.json
 libpldmresponder/test/pdr_jsons/state_sensor/malformed/sensor_pdr.json
