+++ libpldmresponder/meson.build
@@ -27,7 +27,6 @@ sources = [
     '../host-bmc/host_pdr_handler.cpp',
     '../host-bmc/utils.cpp',
     '../host-bmc/dbus_to_event_handler.cpp',
-    '../host-bmc/dbus_to_terminus_effecters.cpp',
     '../host-bmc/host_associations_parser.cpp',
     '../host-bmc/host_condition.cpp',
     '../host-bmc/dbus/asset.cpp',
