+++ platform-mc/test/host_effecter_jsons/good/dbus_to_terminus_effecter.json
