+++ platform-mc/test/host_effecter_jsons/malformed/dbus_to_terminus_effecter.json
